refactor: :recycle: reorganize example tests and roundtrips
